fix book api
