--- pages/book-list/book-list.json
@@ -1,3 +1,0 @@
-{
-  "usingComponents": {}
-}
