+++ pages/user-page-detail/record/record.json
@@ -1,3 +1,6 @@
 {
-  "usingComponents": {}
+  "usingComponents": {},
+  "navigationBarBackgroundColor": "#FFFFFF",
+  "navigationBarTitleText": "",
+  "navigationBarTextStyle": "black"
 }
