+++ pages/user-page-detail/record/record.wxml
@@ -1,2 +1,17 @@
-<!--pages/user-page-detail/record/record.wxml-->
-<text>pages/user-page-detail/record/record.wxml</text>
+<view class="main">
+  <view class='weui-form__text-area'>
+    <view class="weui-form__title">预约记录</view>
+  </view>
+  <button class="btn" bindtap="load" type="primary" size="mini">刷新列表</button>
+
+  <block wx:for="{{list}}" wx:key="index">
+    <view class="view_tupian_wenzi" bindtap="click" id="{{index}}">
+      <view class="view_wenzi2">
+        <text>预约时间：{{item.date}}</text>
+        <text class="text_small">人数：{{item.numPeople}}</text>
+        <text class="text_small">状态：{{item.state}}</text>
+      </view>
+    </view>
+  </block>
+  <view class="text_last">-------------到底了哦 :) -------------</view>
+</view>
