+++ pages/user-page-detail/record/record.wxss
@@ -1,1 +1,61 @@
-/* pages/user-page-detail/record/record.wxss */
+/* pages/book-list/book-list.wxss */
+
+.main {
+  margin: 30rpx;
+}
+
+.main .title1 {
+  font-size: 40rpx;
+  font-weight: bold;
+  height: 60rpx;
+  line-height: 60rpx;
+  padding-left: 30rpx;
+  color: white;
+  border-radius: 10rpx;
+  background-color: #ff6f3b;
+}
+
+.view_tupian_wenzi {
+  height: 120rpx;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  border-radius: 20rpx;
+  border: 3rpx solid #E0E3DA;
+  box-shadow:5rpx 5rpx 5rpx 5rpx #E0E3DA;
+  background-color: #ffffff;
+  margin: 30rpx;
+  padding: 30rpx;
+}
+
+.image_1 {
+  width: 100rpx;
+  height: 100rpx;
+}
+
+.view_wenzi2 {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+}
+
+.text_small {
+  font-size: 30rpx;
+  word-break: break-all;
+  color: #7a7878;
+  margin-top: 10rpx
+}
+
+.text_last {
+  text-align: center;
+  margin-top: 60rpx;
+  font-size: 30rpx;
+  word-break: break-all;
+  color: #7a7878;
+}
+
+.btn {
+  left: 30%;
+  margin: 40rpx auto;
+  width: 40%;
+}
